Refactor: 날짜 형식 변경 함수 utills 폴더로 관리, folders -> links로 명칭 통일
