+++ src/utills/formatDate.js
@@ -1,0 +1,6 @@
+function formatDate(value) {
+  const date = new Date(value);
+  return `${date.getFullYear()}. ${date.getMonth() + 1}. ${date.getDate()}`;
+}
+
+export default formatDate;
